requirements.txt
pymongo==4.6.1
jinja2==3.1.3
openai==1.12.0
python-dotenv==1.0.1
flask==3.0.0
streamlit==1.29.0
